Add FLEX Protocol Support
This commit adds support for the FLEX protocol for pagers, using the
public-domain, single-header library 'tinyflex'.

The tinyflex library currently supports:

- 2-FSK modulation at 1600bps
- Short and long capcodes
- Alphanumeric messages up to 248 characters

The FLEX protocol is a streaming protocol, meaning it requires a
continuous data stream to maintain synchronization with the pager.
However, it is possible to force re-synchronization by sending a
preliminary ERS (Emergency Re-synchronization). This allows
messages to be delivered even if the pager is not actively
synchronized with a transmitter.

This enables behavior similar to the already-supported POCSAG
protocol in rpitx.

Successfully tested on: Raspberry Pi 1B+ (Raspbian 10) + Motorola
Advisor Elite.

Signed-off-by: Davidson Francis <davidsondfgl@gmail.com>